Create readme.md
Hey @JLaferri and @Abrahamh08 ! It's been forever, but my "web-based version" of this is finally live (just in time for SF6)! I thought I'd let you know, and humbly propose this README that mentions the new webapp as an alternative. No worries if you don't accept it, just doing my best to get the word out :)

Happy TO'ing!